+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog 
 =========
 
+.. NEXT
+  - Added Tranform.inverted() to get the inverse transform.
+
 Pymunk 7.1.0 (2025-06-29)
 -------------------------